Enhanced edit capabilities
The edit action now unmarshalls and then compares the Oa4mp server
respresentation of the client with the current data before the edit
to determine if the client has been edited or changed outside of the
plugin. If it detects that the current data and the Oa4mp server
configuration are not synchronized it will not perform the edit and
displays an error that the user should mail help@cilogon.org for
assistance.