Warn on standard claims and check scopes
If the user enters a standard claim into the OIDC Claim Name for a LDAP
to Claim mapping warn the user. Also if a standard claim is entered but
the required scope is not checked put up a dialog box to ask the user if
the scope should be checked.