Different exec_phase in cfg
Write the JSON cfg object such that the exec_phase configuration
for both QDL Scripts includes post_auth, post_token, post_refresh,
and post_user_info. This change coordinates with a change in
the QDL so that an LDAP claim source is created and used directly
to get claims rather then returning the LDAP claim source to the
OA4MP core for it to query LDAP. This change means that not
finding the record in LDAP is no longer an error and the processing
continues so that just the claims from the upstream IdP are
returned.